requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
anthropic>=0.40.0
python-dotenv>=1.0.0
